Generalization of Verlet list which use an underlying linked cells container
