+++ src/autopas/containers/VerletClusterLists.h
@@ -7,6 +7,7 @@
 #pragma once
 
 #include <cmath>
+#include "autopas/cells/FullParticleCell.h"
 #include "autopas/containers/ParticleContainer.h"
 #include "autopas/utils/ArrayMath.h"
